requirements.txt
streamlit>=1.27
fastapi
uvicorn
requests
httpx
numpy
pandas
statsmodels
pywavelets
scikit-learn
torch
aiosqlite
pyyaml
pydantic
altair
altair_saver
pip-audit
plotly
pytest
pytest-asyncio
pyttsx3
Pillow>=9.0.0
matplotlib
keyring
